Moved parameter management code out to separate module
